moved agent diff calculation to agents and cleaned up app.js
